+++ src/modules/mo-product-dto/product/data-transfer-objects/organization/organization-user-group-create-request.dto.ts
@@ -1,5 +1,5 @@
 import { Expose, Type } from 'class-transformer';
-import { IdDto } from '@mo/core-dto';
+import { IdDto } from '@mogenius/core-dto';
 import { IsNotEmpty } from 'class-validator';
 
 export class OrganizationUserGroupCreateRequestDto {