+++ src/modules/mo-project-dto/data-transfer-objects/project-namespace-service-pod/deprecated/project-namespace-service-pod-flat.dto.ts
@@ -1,5 +1,5 @@
 // import { Expose, Type } from 'class-transformer';
-// import { BaseEntityDto } from '@mo/database-dto';
+// import { BaseEntityDto } from '@mogenius/database-dto';
 // import { KubernetesEventDto, KubernetesEventStatusReasonEnum } from '../../../mo-kubernetes';
 //
 // export class ProjectNamespaceServicePodFlatDto extends BaseEntityDto {