+++ src/modules/mo-project-dto/data-transfer-objects/stats/pod-metrics.dto.ts
@@ -1,0 +1,11 @@
+import { Expose, Type } from 'class-transformer';
+import { ContainerMetricsDto } from './container-metrics.dto';
+
+export class PodMetricsDto {
+  @Expose()
+  name: string;
+
+  @Expose()
+  @Type(() => ContainerMetricsDto)
+  containers: ContainerMetricsDto[];
+}